Move the background code into it's own object.
Right now I'm not passing in the background and using a static one. Might want
to change that if I decide to have multiple backgrounds in the future. Also, the
background moves with a constant speed when the character moves and takes dt
into no account, should probably update this to not only take dt into account but
to move the background faster when the character is on the extremes so it looks
like the character is moving at a constant rate.

Adding game.love to gitignore because it's officially too big for github.

Deleting game.love because it's huge.